+++ superstructure/generic/Export.F90
@@ -2,10 +2,11 @@
 ! Public API exposed to external consumers.
 module mapl_generic_export
 
-   use mapl_UserSetServices_mod, only: user_setservices, AbstractUserSetServices, DSOSetServices
+   use mapl_UserSetServices_mod, only: user_setservices
 
    implicit none
    private
 
+   public :: user_setservices
 
 end module mapl_generic_export