+++ superstructure/generic/transforms/NormalizationTransform.F90
@@ -212,7 +212,7 @@ contains
 
        ! TODO (future): When coupler support is added, run coupler to update aux field
        ! if (associated(this%aux_coupler)) then
-       !    call this%aux_coupler%run(phase_idx=GENERIC_COUPLER_UPDATE, _RC)
+       !    call this%aux_coupler%run(phase_idx=MAPL_GENERIC_COUPLER_UPDATE, _RC)
        ! end if
 
        ! Get the field or bundle to normalize