+++ superstructure/generic/vertical/ModelVerticalGrid.F90
@@ -5,6 +5,7 @@ module mapl_ModelVerticalGrid_mod
    use mapl_ErrorHandling_mod
    use mapl_KeywordEnforcer_mod
    use mapl_vertical_grid_export
+   use mapl_BasicVerticalGrid_mod, only: BasicVerticalGrid
    use mapl_field_export
    use mapl_StateRegistry_mod
    use mapl_VirtualConnectionPt_mod